# other/openapi.yaml
openapi: 3.0.0
info:
  title: Yelp API
  description: Yelp API for CMPE 202
  contact:
    name: Group name
    url: http://localhost:8080
  version: 1.0.0
servers: 
  - url: http://localhost:8080
paths: 
...
  /user/{username}:
    get:
      tags:
        - user
      summary: getUserByName
      description: Get user by name
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'          
        '400':
          description: Invalid username supplied
    put:
      tags:
        - user
      summary: updateUser
      description: Update user info
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: name that need to be deleted
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid username supplied
    delete:
      tags:
        - user
      summary: deleteUser
      description: Delete user.
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid username supplied
...
  /restaurant/{restaurantname}:
    get:
      tags:
        - restaurant
      summary: getRestaurant
      description: Get Restaurant
      operationId: getRestaurant
      parameters:
        - name: restaurant
          in: path
          description: 'The restaurant that needs to be fetched. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'          
        '400':
          description: Invalid restaurant supplied
    put:
      tags:
        - restaurant
      summary: updateRestaurant
      description: Update restaurant info
      operationId: updateRestaurant
      parameters:
        - name: restaurant
          in: path
          description: Restaurant that need to be deleted
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent restaurant
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restaurant'
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid restaurant supplied
    delete:
      tags:
        - restaurant
      summary: deleteRestaurant
      description: Delete restaurant.
      operationId: deleteRestaurant
      parameters:
        - name: restaurant
          in: path
          description: The restaurant that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid restaurant supplied
  /search/{query}:
    get:
      tags:
        - search
      summary: getQuery
      description: Get user query
      operationId: getQuery
      parameters:
        - name: query
          in: path
          description: 'The query that needs to be fetched. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'          
        '400':
          description: Invalid query
    put:
      tags:
        - search
      summary: updateQuery
      description: Updates user query
      operationId: updateQuery
      parameters:
        - name: query
          in: path
          description: 'The query that needs to be fetched. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'          
        '400':
          description: Invalid query
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: theUser
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        userRole:
          type: integer
          description: User Status
          format: int32
          example: 1
    Restaurant:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: restaurant
        category:
          type: string
          example: Fast Food
        rating:
          type: integer
          format: int64
          example: 10
        price:
          type: integer
          format: int64
          example: 10
        address:
          type: string
          example: 1234 Street Ave
        zipCode:
          type: integer
          format: int64
          example: 12345
        description:
          type: string
          example: Fast Food
        photos:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/Photo'
              - $ref: '#/components/schemas/Photo'
        hours:
          type: integer
          format: int64
          example: 00
        ownerId:
          type: integer
          format: int64
          example: 12
...
    Photo:
      type: object
      properties:
        url:
          type: string
          example: "String"